Fix #109 (nova show name not working).
By default when searching via name we will do a findall(name=blah) and
due a REST /details which is not the same as a .get() and doesn't get
the information about flavors and images. This fix it as we redo the
call with the id which does a .get() to get all informations.