Add ability to force debugging via os environ.
Currently we can do debug from the command line or integratng to an
application which has already logging configured. But if we want to do
it to experiment via ipython the process is quite cumbersome. This allow
to have an environement variable called NOVACLIENT_DEBUG which enable
debugging when set and defined.