Add MANIFEST.in and setup.cfg back.
These files are actually needed - without MANIFEST.in, files are missing
from the source tarball which is used in debian packaging. setup.cfg is
needed for configuration of the sphinx doc build commands.

Additionally, I added listings for the other files in the root to
MANIFEST.in so that the source tarball would be complete.

Change-Id: Ia0f281663b58baaf2ee23f627d4b1887d0eddfe9