Fix bug 904364: Consistiently handle trailing '/' on URLs
novaclient used to require a trailing '/' on --auth-url and would
mysteriously break if it was not present.  This is mostly due to
urlparse.urljoin()'s behaviour; it was only used here for concatenation
so it was eliminated and trailing '/' chars consistiently stripped.
All url concatenation now requires the second element to begin
with '/'.

Change-Id: I99e95ea92682d917348201bc3190a1e77bbcbff0