More friendly keyring support when token caching is off.
The token caching would always try to save the token even
if --no_cache was specified. The downside of this is
environments where there is no keyring agent it would prompt
for a password. Now, if --no_cache is specified it won't try
to save the token.

Also ability to reset timings (--timings) which is handy for
client-driven profiling efforts. Yes, this should be a separate
branch, but it's so tiny it's hardly worth the effort.

I'd also like to take this time to express my admiration for
Brian Waldon. A stalwart young fellow and a fine coder.

Change-Id: I97c2c5f6864fe156a3e204927b2831b74fb1c893