Add support for hypervisor-uptime.
Adds support for the new uptime call to the hypervisors extension to
nova.  This allows the results of "uptime" on the hypervisor to be
returned to the caller.  Note that currently, only XenAPI implements
the underlying method; if the version of nova uses another virt
driver, the result will be a 501 error.

Change-Id: I34cc92557e6d22705a5591e54404b32245dbe16a