Fix image-create --poll race-condition.
This fixes a race-condition caused by `do_image_create` with polling
enabled returning after the image has finished uploading but before the
instance `task_state` was cleared out.

This would cause a snapshot taken immediately afterwards to fail because
the `task_state` would still be in the `image_snapshot` state.

The (short-term) fix is to add an addition poll-loop that ensures the
`task_state` field is cleared-out as well.

Change-Id: I3adeed24ecea127c8bdd12143634a6ce5da64330