Makes the OS_NO_CACHE env variable work again
The commit to replace --os-no-cache with --os-no-cache works fine
with the cli options, but the env variable is stored in os_cache
which has the opposite of the intended effect. This patch converts
the variable to a bool and then inverts it before it stores it
in os_cache. This makes it work properly again.

Fixes bug 1089696

Change-Id: Iea12806603ecdc39c6475ad4d6f867ebb1e01633