Make --tenant a required arg for quota-show
Fixes LP 1088519

Previously, novaclient was incorrectly defaulting to the tenant name
as the tenant ID when retrieving quota thresholds via the os-quota-sets
API extension, in an apparent attempt to default to the current tenant
if no specific tenant is explicitly requested.

As a result, the default quotas were always returned, regardless of
whether there were specific overridden quotas for this tenant.

We now require that the --tenant option is always specified for the
quota-show verb, as a sensible default isn't always possible
(because the tenant ID may be opaque to the client, for example when
the auth token is cached in the keyring from a previous call out to
keystone - now the tenant ID is of course generally embedded in the
nova publicURL retrieved from the service catalog, but that is not
guaranteed to be the case, i.e. I don't think we can safely make
assumptions about that URL format).

This change also makes the quota-show verb more consistent with the
quota-update verb, which currently requires that the tenant is
always explicitly specified.

Change-Id: I1706ad993059e70ca0e2f3bcf7b1d06cbcc39f2d