Implement get password for novaclient
This will download and decrypt a base64 encoded encrypted password
from the os-server-password extension. It depends on the user
having openssl installed, but if there is an error of any kind it
will print out the encoded and encrypted password instead.

It also implements clear_password which will delete the password
so it can no longer be retrieved.

Change-Id: I2c4e6c3f03b70dc98d6d339381648a6058f46e21