Allow for image selection using the metadata properties.
Instead of forcing users to remember a image name and/or uuid
instead allow them to specify a image metadata property to 'key'
off of.

For example:

I want to boot a image of ubuntu. This could now be a property
set in the image metadata and the following could be used to
find said image.

$ nova boot my-vm --image-with type=ubuntu

Now that user does not need to care what ubuntu version is being
selected or what the uuid is or what the image name is, they just
know that some image with ubuntu will likely get selected.

Change-Id: Ifaf71e282b8411d2e303cb52fd2623f6cf0bf32a