_get_secgroup returns first group even if multiple groups match
Nova's security group api layer ensures that a given tenant is not able to
create multiple security groups with the same name. That said, if one is the
admin user and does:
"nova secgroup-delete <secgroup that has a duplicate name as another tenant>"
The first security group that is returned with that name will be removed,
which might not have been what the user intended to do. In addition, quantum
allows multiple security groups to be created with the same name therefore
this change is needed to ensure the correct behavior when nova has a quantum
security group proxy.

Fixes bug 1105542

Change-Id: I99734355026e2ad18056c2f4180b5740fda8775a