Fix default format of 'nova coverage-report'
'nova coverage-report' returns an error because nova server returns
error if the format is HTML and combine is False.
And 'nova coverage-start' and 'nova coverage-report' do it now.

After applying this patch, we can avoid this problem and text format
report is created when we use 'nova coverage-report'

Fixes bug 1110972

Change-Id: Ib3d05453de638157152a50ea814e44e6cbc19348