CLI for disable service reason
Adds cli option to allow users to give reason
for service-disable. Also adds disabled reason
as a column in service list, so any disabled service
can be seen with reason.

A recent nova change that allows disable-log-reason
allows users to provide reason for disabling service.
This just adds the cli option for the method.

Blueprint record-reason-for-disabling-service

Change-Id: If263788c34279d6b4c568d5e0320448d2ff67a12