Fix the help text process and the generated wrong help
The wrong help text of bash-completion is generated by the process of
choosing first line from the source code comment. We should get the all
help text from the source code comment, and use it.

So this commit removes this unnecessary process, and also fix the
README.rst for proper help text.

Fixes bug 1206005

Change-Id: Id4f5b6a7722197f09eb366ba8f3e643e1af91805