Support programmatic use of disk config extension
Allow servers to be created, resized, and rebuilt with the disk_config
option:

http://api.openstack.org/api-ref.html#ext-os-disk-config

There is a separate extension that exists for disk config already, but
it only works via the novaclient CLI interface, not via programmatic use
of python-novaclient as a library.

assert changes in tests/v1_1/fakes allow for other data in rebuild
and resize requests.

Change-Id: I8051ffb8747cf5c67b0199d22fbbe80306b01499