Add support for os-assisted-volume-snapshots
This patch adds support for the assisted volume snapshots API extension.
This is used by Cinder to ask Nova to perform a volume snapshot on its
behalf.  It's required when the volume is actually file backed (like
qcow2) and the hypervisor needs to be involved in the snapshot
operation.

Required for blueprint qemu-assisted-snapshots

Change-Id: I50ee9bf92c8de98528638d1724fe35e07bed729e