Add-in some re-auth logic when using os_cache
When using os_cache for the auth_token, if username and password
are available when a token fails to validate/authenticate, attempt
a reauth behind the scenes.

The helper object for stashing the auth_token doesn't handle flushing
the cache.  This is valid behavior because the result would be the
same as if a bad username/password were issued (401 Unauthorized).

Fixes-bug: 1214658
Change-Id: I107fc999610bb7638cefcb5d195aeafdcd663ca6