Make os-cache retry on an invalid token
client.authenticate doesn't actually do any rest calls if a token is present,
it just assumes the token is valid.  Instead re-authentication is done if the
request raises an Unauthorized exception, at which point we want to
unauthenticate (delete the token from in-memory) and overwrite the old token
in the keyring. Also if a password is present allways pass it into the client
so re-authentication is possible during _cs_request.

Change-Id: I86d0356d5685ffa802f1bdc97e33727ff2dd075e