Remove tox locale overrides
* tox.ini: The LANG and LANGUAGE environment overrides were
introduced originally during the testr migration in an attempt to be
conservative about the possibility that locale settings in the
calling environment could cause consistency problems for test runs.
In actuality, this should be unnecessary and any place where it does
cause issues ought to be considered an actual bug.

Change-Id: I0f30b45033145c61ad2a9b232cb1178c48c7fc97