Removes copy of output of 'nova help' from README
Removes the copy of the otuput of 'nova help' from the README.rst
file. It's out of date and given its not autogenerated is
rather prone to getting out of date again soon anyway. Just leave
the command that people need to run in order to print the help message.

Change-Id: I3cd069838302913bc69cc0ded5d632c7f8c3f408