Allow admin user to get all tenant's floating IPs
When getting floatingips with Nova API, the results will be filtered
with the 'tenant_id'.
So, we can only get the floatingips belonging to the tenant of the current
context.
When ceilometer invokes novaclient to list floatingips, it will get an
empty list because the tenant is 'service'.
we should allow an admin user to index all tenants's floatingip by adding a
parameter 'all_tenants'.
This patch provides CLI support

Change-Id: I35a2155401247d49017bf3c03341b082cb87750d
Closes-bug: #1262124