return 130 for keyboard interrupt
when keyboard interrupt is received by novaclient, the return code as
of now is 1.

But since the client was terminated by an keyboard interrupt, the return
code should be 130. (http://tldp.org/LDP/abs/html/exitcodes.html)
It is useful when people are writing automation test cases and want to
validate based on the return code.
I have also changed the message that is printed on shell to match other
clients.

Change-Id: If544ade154c53b1f18518773f4d49cd335a394d0
Closes-Bug: #1373231