Pass kwargs through to the adapter
A major point of the adapter is that we should be able to add new
arguments to the adapter that are supported automatically by the clients
that support it. This means we have to pass all additional arguments
through to adapter.

Regardless of whether the arguments are passed to the adapter or to the
traditional client the kwargs are explicitly provided so we won't end up
with missing arguments.

Closes-Bug: #1403329
Change-Id: I7303b184875dc296ca855ad89496173c42070879