Add Client object to documentation
After wanting to do a rather simple thing -- figure out of the
"timeout" argument to nodepool was a int or a float, it lead me down
the rabbit-hole to python-novaclient.

It turns out the timeout does get passed through to requests as a
float so that mystery was solved.

But the "Client" class seems to be missing from the documentation as
it's not included in the class list.  So add that and also at least
document the types of the arguments.

However, then I noticd that this wasn't showing up; turns out sphinx
requires "autoclass_content = both" if you want it to document
__init__() functions.  Several other classes had their init args
documented but they weren't showing up because of this.

Change-Id: I8f44e92f2a0f25a75926b1813a8b374e79b4f5db