Add missing servers.create parameter documentation
The min_count, max_count, and security_groups parameters for
servers.create are missing documentation. The security_groups
parameter was documented in the servers._boot function but that
private function isn't published to the docs.

This change adds the missing documentation for the aforementioned
parameters and removes the duplicate docs in the servers._boot
function.

Change-Id: Ib0dcc7e8f7f4d0b83b6caeddd1ace66add4e0003