refactor functional test base class to no inherit from tempest_lib
Test base classes inheriting outside the existing source tree to
anything higher up the stack than testtools is really an anti
pattern. It makes it *far* less clear what is going on. We really need
to own and understand our base class setup for tests.

This does that unwind, so that we only now call out to tempest_lib in
specific ways (like building clients, using decorators). The timeout
and log capture pieces are pulled inline.

At the end of this we end up with a base test that defines:

 - self.client - a Nova API client
 - self.flavor - a workable flavor for booting guests
 - self.image - a workable image for booting guests
 - self.cli_clients - tempest_lib cli clients

Change-Id: I716be51d7d1825a757934298f06b2f04d64cf0dd