fix FloatingIP repr
The FloatingIP repr added in d614dbcab9cc8c5732ce28c655738218e147aad7
is incorrect, and can cause fails when you attempt to do bulk creation
activities with logging turned on.

This removes the __repr__ for FloatingIP so that it fails back to the
Resource defaults which are safe.

Partial Revert of d614dbcab9cc8c5732ce28c655738218e147aad7

Closes-Bug: #1437244

Change-Id: I8cedf418917157ce632606bef640049140134372