Fix resolving image.id in servers.boot
Most places resolve id of input parameter as 'base.getid'
and in the beginning of _boot function it resolves image id
with such method. But in fixed place it tries to get id directly
from object. But code that uses novaclient directly(without shell)
falls in this place because it passes string with id to image parameter.

Change-Id: Ib90e4ffa3b7835f6648a62eddf7f7b87b5ab0f7e