fix novaclient functional tests for new devstack config
The novaclient functional tests assumed computev21 as a type in the
service catalog. We're defaulting compute in devstack to v21 now, so
this no longer works. This is probably far too coupled to devstack
atm, however until we get to unversioned service catalog, this is what
we need to do.

Change-Id: Ifb0056924b20f62d75d13a7b4aae829df61a5396