Fix H404/405 violations in novaclient/v2/[a-f]
There is a lot of H404/405 violations in novaclient, and it is better
to fix those to have a better doc string for class/methods.

This patch fixes these violations for [a-f] files under  novaclient/v2
folder.

As there are lot of violations and cannot be fixed in single patches,
So separating those in multiple patches for easy review.

Change-Id: Ibe5afef349d1be6d70903e06bee402ca30738ea1
Partial-Bug: #1521899