Fix host-evacuate-live for 2.25 microversion
Change I01b22593724616bc0a7793c509ecabf095d6927d made the
live_migrate() method in the ServerManager conditional on the
API version requested. This broke the host-evacuate-live
command which is calling ServerManager.live_migrate() directly
with one too many arguments for the v2.25 version of the method.

This updates the host-evacuate-live shell to behave like the
live-migration method and be aware of the API version when
calling the live_migrate() method.

Related to blueprint making-live-migration-api-friendly

Conflicts:
        novaclient/v2/contrib/host_evacuate_live.py

NOTE(mriedem): The conflict is due to
93913c84b67d3aefe79bef106fbe9a83f0bcdb71 not being in stable/mitaka.
Also note that this is not for a blueprint, it's just related
to a blueprint that landed in mitaka but introduced a regression
here when support was added for that blueprint with the 2.25
microversion.

Change-Id: I4dbeb6ebe03f03799b706be2d787d21484b5c664
Closes-Bug: #1561938
(cherry picked from commit 90fbbb29562a905e8f70badf2c31cfb4ec6841ed)