Update to microversion 2.27
This enables the use of the OpenStack-API-Version header, in
addition to the older X-OpenStack-Nova-API-Version header.

If the client knows that it is in a version less than 2.27, it
will not send the newer header.

If the client knows that is in a version >= 2.27 if it gets a
response that does not have the expected response header (the
newer one) it will warn. If it is in an older version, it will
warn about the other one missing.

A server that is 2.27 or beyond will accept both headers and
respond with both.

Change-Id: I1f7b1ca0fe795e4ecd333de761d96fff117969c0