Use neutron for network name -> id resolution
The 2.36 microversion deprecates the nova proxy APIs for looking up a
network from neutron. The CLI will lookup a network by name for
booting a server as a convenience.

Over 90% of clouds now have neutron. We use a service catalog lookup
to assume that neutron is available, and if not fall back to our
proxy. This should mostly shield people from the 2.36 transition.

To test this we add support in our fake client for the neutron network
lookup. Comments were also left about some cleanups we should do in
other parts of the shell testing for net lookup.

Related to blueprint deprecate-api-proxies

Change-Id: I4f809de4f7efb913c814f132f5b3482731c588c5