Fix test_trigger_crash_dump_in_locked_state_nonadmin test
Patch 2a70e9688151588347664badbac5d0a59ef0f95c in the server
changed the error message that comes back in the case of a locked
server which breaks this test. The test is unnecessarily rigid in
it's assertion, we should just be checking that we get a 409
back, not what the actual error message is, which is prone to change.

Change-Id: I1fd64953e671c6c7fe1068a9c30f5a077712a9fb
Closes-Bug: #1618561