Fix functional tests to deal with multiple networks
There was a change in the openstacksdk 0.9.11 which
python-openstackclient uses to create networks in a devstack
run with Neutron. Because of this change, the admin tenant
has access to both the 'public' and 'private' network setup
in devstack which before used to just be the 'public' network,
which exposed a bug in the novaclient functional testing where
the admin user is attempting to create a server but not
specify a specific network to use, but multiple networks
are available to the admin (it can list multiple networks).
In this case the networks are the standard public and private
networks that are created in devstack. Since a network isn't
specified when creating the server, the nova API fails with a
409 error because it can't determine which network to use.

This patch fixes the testing in novaclient by checking to see
if there are multiple networks available and if so, specifies
one for the legacy BDM tests that weren't specifying a network
ID (those tests don't really care about the networking).

The auto-network test is skipped if there are multiple networks
available because passing in a specific network would defeat
the purpose of that test.

Change-Id: I22ee148581a94b153cf7e733563cfafaa56b1ffd
Closes-Bug: #1654806