Microversion 2.42 - Fix tag attribute disappearing
The 2.32 microversion added the support for creating a
server with a tagged block device mapping and/or tagged
nic.

Due to a bug, block device tags would only work in the
2.32 microversion but nic tags would still work up until
the 2.37 microversion, which regressed that functionality
as well.

Both block device and nic tags were fixed again in the 2.42
microversion.

This change updates the help text for the 'nova boot' CLI
and client-side validation in the shell for the various
microversion ranges of support for both block device and
nic tags.

Co-Authored-By: Andrey Kurilin <akurilin@mirantis.com>
Change-Id: I7492b20b5d2daa0f8391abed07ef861043c3e51e
Implements: blueprint fix-tag-attribute-disappearing
Closes-Bug: #1658571