2.47: Show flavor info in server details
This adds support for microversion 2.47 which directly embeds the
flavor information in the server details.  With this change, CLI
requests with microversion >= 2.47 will no longer need to do
additional queries to get the flavor and flavor extra_specs
information.  Instead, the flavor information will be output
as separate key/value pairs with the keys namespaced with the
"flavor:" prefix.  As one would expect, these keys can also be
specified as output fields when listing servers.

Change-Id: Ic00ec95485485dff0fd4dcf8cad6ca56a481d512
Implements: blueprint instance-flavor-api
Depends-On: If646149efb7eec8c90bf7d07c39ff4c495349941