Add functional test for "nova boot --image-with"
We have no functional testing for creating a server
using the --image-with option, which is going to be
impacted when we remove our deprecated code for the
image proxy API in Nova.

This adds a functional test for that flow which
creates a server, snapshots the server with a meta
key/value, then creates a second server using
--image-with and specifying the same meta key/value
set on the snapshot image.

Change-Id: I30017380d3e8dba6f1a50dfdda5eb4802b498bb4