Remove deprecated network-related resource commands
In Newton we deprecated the various network-related resource
commands which were either for nova-network or were proxy
APIs to Neutron, because of the 2.36 microversion which deprecated
those APIs.

This removes the deprecated commands. Because of the size of this
change, the deprecated python API binding removals will come in a
separate change.

Change-Id: I6ecca49e7208f9dc0969bf377930a729c4b407b8