Remove direct dependency on requests
keystoneauth is the library that pulls in requests. If we have client
libraries also depend directly on requests, we can get into situations
for end users in installations where, because of sequencing of client
library releases there are conflicting versions of requests needed and
pkg_resources/entrypoints break.

Remove the direct depend and let keystoneauth pull it in for us.

Change-Id: I76fe89f9c166e61971b5c116c7f01fbd5ccb5cf7