Fix help message and description for volume-update
The help for the `volume-update` command says:

update_server_volume(server_id, attachment_id, new_volume_id)
Update the volume identified by the attachment ID, that is
attached to the given server ID

Parameters:
server_id – The ID of the server
attachment_id – The ID of the attachment
new_volume_id – The ID of the new volume to attach
Return type:
Volume

But in reality the command takes an original volume and a
destination volume. The update help is also not very clear
regarding what the command actually does and the fact that
it's primarily a libvirt data migration from src volume to
destination volume.

This patch modifies the help to show the correct arguments
and also adds some more info regarding what the command
is doing.

Change-Id: Ia267285f85854976545c2a72c3729c1bfcb2c991
Closes-Bug: #1690397