Prevent 2.32 release note from showing up in 9.0.0 release notes
The release note for the 2.32 microversion support was originally
added in the Newton release and tagged on 5.0.0:

7fb0bd4f357a7f9e24c864584b398a25171580cf

Then it was modified on master (pike) with change:

aa7f35d76b5768ec3fe4d45ad010d53adff7cccd

That makes reno think that it's a new unreleased release note,
which makes it show up in the unreleased notes which are going
to eventually be the release notes for the 9.0.0 release. This
is wrong and confusing since 2.32 support happened in 5.0.0 not
9.0.0.

Reverting the typo change doesn't help since reno is still picking
up the release note from that change in the git history. The only
way to remove the release note from being processed on the current
branch (master) is to delete it. The note would still be available
in stable/ocata and stable/newton branches.

Change-Id: Ie6e9b930ff49621244df286d8f1d2cedffa2ff11