Fix setting 'global_request_id' in SessionClient
In Ic75be3acb8b77aae8da631e3c4cd6f545a9a35cb,
'global_request_id' is set
in the constructor of class keystoneauth1.adaptor.Adapter.

But 'global_request_id' was not passed to the constructor,
'global_request_id' was cleared.
It caused the test failure.

It is not necessary to set 'global_request_id'
in the constructor of class SessionClient
and just passing 'global_request_id' in kwargs
is required.

Change-Id: Id587e35c221fe2b11889469f88557d254125ea7e
Closes-Bug: #1697358