[ci] Use pseudo-random names for new resources
Random names of resources makes the task impossible to find a test which
doesn't clean resources after a launch.

This patch changes name_generate method to generate names including a name
of the test, which can be really helpful.

Also, the simple script that prints resources after test run is added.

Change-Id: Id3a743afb624dd4bf8ed3523a1916455fd93655a