Fix cropping the endpoint url
`nova version-list` crops the endpoint url to exclude API version and
project-id(for microversions < 2.18). It is ok, but the code was
designed before Nova-API moved under uwsgi, which changed the endpoint
url format.

The latest endpoint url contains the hostname and port which can be one
for several services and it is a wrong case to cut the whole path from
the endpoint.

Closes-Bug: #1702194

Change-Id: Icba858b496855e2ffd71b35168e8057b28236119