Change Service repr to use self.id always
Before this change, the Service object repr was
the binary for microversion < 2.53. With microversion
>= 2.53, the Service repr became the id, which is a UUID.

Using the binary never really made sense since if you have
multiple nova-compute services, the binary is going to be
the same for all of them in the repr and nothing is
distinguishable.

This changes the Service repr to just use the id, which is
going to be the integer id value if microversion < 2.53 and
the UUID id value if microversion >= 2.53.

There is no release note for this change since the repr should
not be treated as a contractual API.

Change-Id: I3a7de2683e339295022efb279828ab1a91b3b62e