Be clear about hypevisors.search used in a few CLIs
The following CLIs are doing a pattern match on the hypervisor
hostname when getting servers:

1. nova host-evacuate
2. nova host-evacuate-live (terrible name)
3. nova host-servers-migrate
4. nova host-meta (terrible name)

The fact that the hypervisor host(s) are looked up using
a pattern match is not clear in the help string for the host
argument.

This makes the help more clear and adds a warning about being
specific if you only want to target a specific host for these
operations.

In a later change we may modify the behavior in these CLIs which
allow you to find hypervisors via pattern matching, but for now
we should at least update the help text so we can backport it.

Change-Id: Icec03326bb4d9f898c04e10199038167ce5e3cce
Partial-Bug: #1667794