Remove SecretsHelper
This code hasn't worked for over two years when
9ed9ab68f7d2fb98c505afc3a1079a72f15c959c was added
which does API version discovery, which requires auth,
which happens before the SecretsHelper is used. That was
back in the 2.27.0 release in the Liberty series.

Given how long this has been broken without anyone noticing,
and it's undocumented and untested (basically), and
python-openstackclient properly handles prompting for a password,
we should just remove this code rather than try to fix it.

Change-Id: I62188e73a48f6878ce920a3b4724dba101564aef
Closes-Bug: #1732744