Optimize jobs run on novaclient
novaclient run 2 jobs for functional tests
- novaclient-dsvm-functional-identity-v3-only	(non-voting)
- novaclient-dsvm-functional-neutron

These 2 were added when neturon and identity v3 were not default
in devstack. Now both run as default and we do not separate job
to run.

This commit does below changes:
- delete the 'novaclient-dsvm-functional-neutron'
- In 'novaclient-dsvm-functional-identity-v3-only'
  - make 'novaclient-dsvm-functional-identity-v3-only' as voting
  - cleanup some 'if' condition which are hardcoded true now
  - rename 'novaclient-dsvm-functional-identity-v3-only' to 'novaclient-dsvm-functional'

NOTE: this not going to backport to stable branch as they seems running
identity v2 and v3 in those jobs. Let's keep the same setup there.

Change-Id: I4bc564e548876ef4d3b30e736c0055f19c062319