nova limits ERROR (Exception): Field names must be unique
Running `nova absolute-limits` or `nova limits` can fail
when unaccounted for limits are returned

This works OK for a standard Vanilla OpenStack deployment
today, but it is quite fragile if new limit types are
introduced, or if a deployer sends back
other custom limits.
For instance, my limits include `maxTotalPrivateNetworks`
and `totalPrivateNetworksUsed` limits, which do not map
accordingly, and thus result in an error.

This happens when 'Others' field(custom limit type) is
occuring more than once which is not handled in code
causing multiple columns having same name.
Which throws exception and does not pretty prints
absolute limit.

This fix should also fix any future custom limits.

Closes-bug: #1546767

Change-Id: I1d3cf707722fc71c20cf4ec517b3f4f4875480e0