Remove incorrect legacy QuotaSet.id property
The QuotaSet.id property was added in commit
d17505462db3876a781016b16464e8ed967ae50d assertin
that the os-quota-set response had no id in it, but
it always has, ever since the API was added in commit
6220c4276e30c633ffc4165ce6db0d120c0e88a7 in nova.

This change removes the incorrect property and fixture
data that returns a tenant_id in the response which the
API doesn't actually do.

The test_quota fixture data is also updated to fix the
name on the key_pairs key and add missing keys in the
response for server groups.

Change-Id: Ifaf59813e75876334dcc2ac239ed6bdddb495aa7