Remove deprecated services binary CLI arg
The services CLI 'binary' arg was deprecated in Pike
via change Idd0d2be960ca0ed59097c10c931da47a1a3e66fb
because with cells v2 support in the API, the binary
argument for the enable/disable CLIs doesn't make sense
as the only binary those work with is 'nova-compute'.
So this removes the deprecated argument and hard-codes
the value to be 'nova-compute'.

Change-Id: I60490f3d74212bb172dccc1c7d337198e1021236