Fix AttributeError in getting a resource ID
In the 'wait_for_resource_delete' method of the 'ClientTestBase' class,
the following statement is executed when timeout.

  self.fail("The resource '%s' still exists." % resource.id)

The 'resource' variable is passed in the argument of
the 'wait_for_resource_delete' method.
It is sometimes a string which is a resource ID.

When it is a string (a resource ID), AttributeError is raised
at the statement.
This patch fixes the issue.

Change-Id: I558026de54a5cc75359225b50054934a220f3e5c
Closes-Bug: #1704132