Stop printing flavor details on successful flavor-delete
It's weird that the flavor-delete CLI would print out the
details of the flavor the user just deleted, rather than give
some confirmation message, or simply not print anything out
at all, which is arguably more expected behavior when deleting
a resource. This change simply removes the output on successful
deletion which brings "nova flavor-delete" more inline with other
delete CLIs, including "openstack flavor delete" behavior.

While in here, the variable and docstring code is updated to
reflect reality and a release note is added in case anyone was
ever relying on this odd behavior.

Change-Id: Ie3e07b45cc70213cde4e6077c604c9f2835c75ad
Closes-Bug: #1764420