Import nova CLI reference from openstack-manuals
This replaces the old CLI reference main page for
the various nova commands with the more complete
CLI reference that was in the openstack-manuals
report based on the "before-migration" tag in
that repo, which was the tag before the centralized
docs were dropped in Pike and should have been
moved into per-project repos, in this case
python-novaclient.

The command reference in here is a bit old and
should be cleaned up, but this is better than
what we have today, which is nothing.

Change-Id: If19cfcafc90fddb930c124a5b9845d4eae1f6093
Closes-Bug: #1775281