Remove unnecessary if statement
The interface_list method of
the novaclient.v2.server.Server class always returns
an object of the novaclient.base.ListWithMeta class
which is a sub class of list.

So 'if' statement that check whether it is an instance of list
is not necessary when printing the return value of the method.

Change-Id: I5dc5bfc6a783bb59c3aec2cce626bb00de633754